requirements.txt
# common wheelhouse for ALL images (you can copy-paste into every Dockerfile)
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1

langchain>=0.2.2
langchain-community>=0.2.2
langchain-ollama==0.1.0
sentence-transformers==2.7.0
neo4j==5.20.0
PyMuPDF==1.24.4
pillow==10.4.0
tiktoken==0.6.0

# image embeddings
torch==2.3.0
clip-anytorch==2.5.2

# vector store
chromadb==0.5.0
